task done

+++ razor-main/BadNews/Repositories/Weather/OpenWeatherForecast.cs
@@ -18,10 +18,10 @@ namespace BadNews.Repositories.Weather
 
         public class MainInfo
         {
-            public int Temp { get; set; }
-            public decimal FeelsLike { get; set; }
-            public int TempMin { get; set; }
-            public int TempMax { get; set; }
+            public double Temp { get; set; }
+            public double FeelsLike { get; set; }
+            public double TempMin { get; set; }
+            public double TempMax { get; set; }
             public int Pressure { get; set; }
             public int Humidity { get; set; }
         }
